\name{community_synchrony-package}
\alias{community_synchrony-package}
\alias{community_synchrony}
\docType{package}
\title{
Runs IPMs and other analysis for project on species synchrony in natural communities.
}
\description{
The package contains functions to calculate community synchrony metrics
and to link vital rate regressions to the IPMdoit package for 
simulation tests.
}
\details{
\tabular{ll}{
Package: \tab community_synchrony\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-28\cr
License: \tab MIT\cr
}

}
\author{
Andrew Tredennick

Maintainer: Andrew Tredennick <atredenn@gmail.com>

}
\references{
Chu and Adler 2015, Ecological Monographs
Loreau and de Mazancourt 2008, The American Naturalist
}

\keyword{ package }
\seealso{

}
\examples{

}
