% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crowding_functions.R
\name{estimate_crowding}
\alias{estimate_crowding}
\title{Estimate growth crowding effects on focal species}
\usage{
estimate_crowding(site, data_path, alphas, vital_rate)
}
\arguments{
\item{site}{Name of the focal site (character).}

\item{data_path}{The directory path leading to the data folder,
ending with "/" (character).}

\item{alphas}{The alpha values for each species at the site, in alphabetical
order by species.}

\item{vital_rate}{Vital rate of interest. "growth" or "survival"}
}
\value{
List of crowding matrices, one list entry per species.
}
\description{
Estimate growth crowding effects on focal species
}
\author{
Andrew Tredennick
}

