% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipm_kernel_vitalrate_fxns.R
\name{f}
\alias{f}
\title{Define fecundity function}
\usage{
f(v, u, Rpars, rpa, doSpp)
}
\arguments{
\item{v}{Population vector.}

\item{u}{Population vector copy, essentially.}

\item{Rpars}{List of recruitment regression parameters.}

\item{rpa}{Recruits per area, from get.rpa function.}

\item{doSpp}{Numeric scalar for current species.}
}
\description{
Define fecundity function
}
\author{
Andrew Tredennick
}

