% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import_regression_params_fxns.R
\name{format_recruitment_params}
\alias{format_recruitment_params}
\title{Import and format recruitment parameters for IPM}
\usage{
format_recruitment_params(do_site, species_list, Nyrs, Rdata, path_to_site_data)
}
\arguments{
\item{do_site}{Focal site (character scalar).}

\item{species_list}{Character vector of four letter species' codes.}

\item{Nyrs}{Number of random effects years.}

\item{Rdata}{Recruitment parameters matrix from regression output.}

\item{path_to_site_data}{Directory path to site-specific data folder.}
}
\description{
Import and format recruitment parameters for IPM
}
\author{
Andrew Tredennick
}

