% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import_regression_params_fxns.R
\name{format_survival_params}
\alias{format_survival_params}
\title{Import and format survival parameters for IPM}
\usage{
format_survival_params(do_site, species_list, Nyrs, Sdata)
}
\arguments{
\item{do_site}{Focal site (character scalar).}

\item{species_list}{Character vector of four letter species' codes.}

\item{Nyrs}{Number of random effects years.}

\item{Sdata}{Survival parameters matrix from regression output.}
}
\description{
Import and format survival parameters for IPM
}
\author{
Andrew Tredennick
}

