% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipm_kernel_vitalrate_fxns.R
\name{G}
\alias{G}
\title{Define growth kernel function for IPM}
\usage{
G(v, u, W, Gpars, doYear, doSpp)
}
\arguments{
\item{v}{Population vector.}

\item{u}{Population vector copy, essentially.}

\item{W}{Crowding matrix.}

\item{Gpars}{List of growth parameters.}

\item{doYear}{Current year for getting year level random effects.}

\item{doSpp}{Numeric scalar for species.}
}
\description{
Define growth kernel function for IPM
}
\author{
Andrew Tredennick
}

