% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/community_synchrony.R
\name{get_comm_synchrony}
\alias{get_comm_synchrony}
\title{Calculate synchrony and stability from genet-level time series}
\usage{
get_comm_synchrony(ts_data)
}
\arguments{
\item{ts_data}{Time series dataframe of genet sizes. 'Bad' quad-years should already be removed.}
}
\value{
A list of data frames for synchrony and stability metrics for the community.
}
\description{
Calculate synchrony and stability from genet-level time series
}

