% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/community_synchrony.R
\name{get_ipm_synchrony}
\alias{get_ipm_synchrony}
\title{Calculate synchrony and stability from IPM time series}
\usage{
get_ipm_synchrony(ts_data)
}
\arguments{
\item{ts_data}{A melted dataframe of IPM cover values with three columns: year, species, cover.}
}
\value{
A list of data frames for synchrony and stability metrics for the community.
}
\description{
Calculate synchrony and stability from IPM time series
}

