% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipm_kernel_vitalrate_fxns.R
\name{get_rpa}
\alias{get_rpa}
\title{Define number of recruits per area function.}
\usage{
get_rpa(Rpars, cover, doYear, A)
}
\arguments{
\item{Rpars}{List of recruitment regression parameters.}

\item{cover}{Current time step cover in m^2. Length is num_species.}

\item{doYear}{Numeric scalar for current simulation year.}
}
\description{
Define number of recruits per area function.
}
\author{
Andrew Tredennick
}

