% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vital_rate_regression_functions.R
\name{get_survival_params}
\alias{get_survival_params}
\title{Estimate survival regression coefficients using INLA}
\usage{
get_survival_params(dataframe, crowd_mat, alpha)
}
\arguments{
\item{dataframe}{Time series dataframe of genet sizes at t0 and t1 with
neighborhood crowding covariate for each observation.}

\item{crowd_mat}{Matrix of crowding indices for each observation. Matrix
dimensions are nrow = number of observations,
ncol = number of species.}

\item{alphas}{Vector of length(species) of vital rate specific alpha values.}
}
\value{
Dataframe with named regression coefficients.
}
\description{
Estimate survival regression coefficients using INLA
}
\author{
Andrew Tredennick
}

