% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipm_kernel_vitalrate_fxns.R
\name{S}
\alias{S}
\title{Define survival kernel function for IPM}
\usage{
S(u, W, Spars, doYear, doSpp)
}
\arguments{
\item{u}{Population vector.}

\item{W}{Crowding matrix.}

\item{Spars}{List of survival parameters.}

\item{doYear}{Current year for getting year level random effects.}

\item{doSpp}{Numeric scalar for species.}
}
\description{
Define survival kernel function for IPM
}
\author{
Andrew Tredennick
}

